Begin improving modularization of initialization code
